terminate

+++ par-cube-glucose.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
-/usr/sbin/sshd -D &
 
 if [ -z "$1" ]
 then
@@ -11,12 +10,17 @@ else
   DIR=.
 fi
 
+# check if input file exists, otherwise terminate
+if [ ! -f "$CNF" ]; then echo "c ERROR formula does not exit"; exit 1; fi
+
 PAR=${NUM_PROCESSES}
 OUT=/tmp
 
 if [ -z "$PAR" ]; then PAR=4; fi
 
 echo $PAR
+
+/usr/sbin/sshd -D &
 
 rm $OUT/output*.txt
 touch $OUT/output.txt
@@ -31,18 +35,13 @@ do
   cat $OUT/output*.txt | grep "SAT" | awk '{print $1}' | sort | uniq -c | tr "\n" "\t";
    
   SAT=`cat $OUT/output*.txt | grep "^SAT" | awk '{print $1}' | uniq`
-  if [ "$SAT" == "SAT" ]
-  then
-    echo "DONE"
-    pkill -TERM -P $$
-    FLAG=0
-  fi
+  if [ "$SAT" == "SAT" ]; then echo "ONE JOB SAT"; pkill -TERM -P $$; FLAG=0; fi
 
   UNSAT=`cat $OUT/output*.txt | grep "^UNSAT" | wc |awk '{print $1}'`
   if [ "$OLD" -ne "$UNSAT" ]; then echo $UNSAT $PAR; OLD=$UNSAT; fi
-  if [ "$UNSAT" == "$PAR" ]; then echo "c ALL JOBS UNSAT"; FLAG=0; break; fi
+  if [ "$UNSAT" == "$PAR" ]; then echo "c ALL JOBS UNSAT"; pkill -TERM -P $$; FLAG=0; break; fi
   ALIVE=`ps $$ | wc | awk '{print $1}'`
-  if [ "$ALIVE" == "1" ]; then echo "c PARENT TERMINATED"; FLAG=0; break; fi 
+  if [ "$ALIVE" == "1" ]; then echo "c PARENT TERMINATED"; pkill -TERM -P $$; FLAG=0; break; fi 
   if [ "$FLAG"  == "1" ]; then sleep 1; fi
 done &
